+++ aim/web/ui/src/components/ImagesSet/ImagesSet.d.ts
@@ -1,10 +1,12 @@
+import { IImagesPanelProps } from 'components/ImagesPanel';
+
 import { IFocusedState } from 'types/services/models/metrics/metricsAppModel';
 
 export interface IImageSetProps {
   data: any;
   orderedMap: { [key: string]: any };
-  imagesBlobs: object;
   onScroll: () => void;
+  onListScroll: ({ scrollOffset: number }) => void;
   addUriToList: (blobUrl: string) => void;
   index?: number;
   imagesSetKey: number;
@@ -13,4 +15,9 @@ export interface IImageSetProps {
   imageHeight: number;
   syncHoverState?: (args: any) => void;
   focusedState: IFocusedState;
+  hoveredImageKey: string;
+  setImageFullMode: (v: boolean) => void;
+  setImageFullModeData: (v: string) => void;
+  imageProperties: IImagesPanelProps['imageProperties'];
+  tableHeight: IImagesPanelProps['tableHeight'];
 }