+++ aim/web/ui/src/components/kit/index.ts
@@ -5,7 +5,10 @@ import ToggleButton from './ToggleButton';
 import Icon from './Icon';
 import Badge from './Badge';
 import Dropdown from './Dropdown';
+import JsonViewPopover from './JsonViewPopover';
 import AutoSuggestions from './AutoSuggestions';
+import Slider from './Slider';
+import Menu, { MenuItem } from './Menu';
 
 export {
   Text,
@@ -15,5 +18,9 @@ export {
   Badge,
   Icon,
   Dropdown,
+  JsonViewPopover,
   AutoSuggestions,
+  Slider,
+  Menu,
+  MenuItem,
 };