+++ aim/web/ui/src/config/enums/imageEnums.ts
@@ -1,0 +1,12 @@
+enum ImageRenderingEnum {
+  Smooth = 'smooth',
+  Pixelated = 'pixelated',
+}
+
+enum ImageAlignmentEnum {
+  Original = 'Original',
+  Width = 'Width',
+  Height = 'Height',
+}
+
+export { ImageRenderingEnum, ImageAlignmentEnum };