+++ aim/web/ui/src/pages/ImagesExplore/ImagesExplore.tsx
@@ -190,12 +190,12 @@ function ImagesExplore(): React.FunctionComponentElement<React.ReactNode> {
               onDensityChange={imagesExploreAppModel.onDensityChange}
               getImagesBlobsData={imagesExploreAppModel.getImagesBlobsData}
               imagesData={imagesExploreData?.imagesData}
-              imagesBlobs={imagesExploreData?.imagesBlobs}
               orderedMap={imagesExploreData?.orderedMap}
               isLoading={imagesExploreData?.requestIsPending}
               applyButtonDisabled={imagesExploreData?.applyButtonDisabled}
               panelResizing={panelResizing}
               resizeMode={imagesExploreData?.config?.table.resizeMode}
+              tableHeight={imagesExploreData?.config?.table?.height}
               imageWrapperOffsetHeight={offsetHeight || 0}
               imageWrapperOffsetWidth={offsetWidth || 0}
               isRangePanelShow={
@@ -210,6 +210,9 @@ function ImagesExplore(): React.FunctionComponentElement<React.ReactNode> {
               tooltip={
                 imagesExploreData?.config?.images?.tooltip as IPanelTooltip
               }
+              imageProperties={
+                imagesExploreData?.config?.images?.imageProperties
+              }
               onActivePointChange={imagesExploreAppModel.onActivePointChange}
               controls={
                 <Controls
@@ -219,7 +222,17 @@ function ImagesExplore(): React.FunctionComponentElement<React.ReactNode> {
                   tooltip={
                     imagesExploreData?.config?.images?.tooltip as IPanelTooltip
                   }
+                  imageProperties={
+                    imagesExploreData?.config?.images?.imageProperties
+                  }
                   onChangeTooltip={imagesExploreAppModel?.onChangeTooltip}
+                  onImageSizeChange={imagesExploreAppModel.onImageSizeChange}
+                  onImageRenderingChange={
+                    imagesExploreAppModel.onImageRenderingChange
+                  }
+                  onImageAlignmentChange={
+                    imagesExploreAppModel.onImageAlignmentChange
+                  }
                 />
               }
             />