+++ aim/web/ui/src/services/models/explorer/createAppModel.ts
@@ -517,7 +517,7 @@ function createAppModel({
               );
               const runData = await getRunData(stream);
               updateData(runData);
-            } catch (ex: any) {
+            } catch (ex: Error | any) {
               if (ex.name === 'AbortError') {
                 // Abort Error
               } else {
@@ -624,6 +624,7 @@ function createAppModel({
               epoch: '',
               time: '',
               children: [],
+              groups: groupConfigData,
             };
 
             rows[groupKey!] = {
@@ -786,7 +787,7 @@ function createAppModel({
               rows[groupKey!].data,
               {},
               true,
-              ['value'].concat(Object.keys(columnsValues)),
+              ['value', 'groups'].concat(Object.keys(columnsValues)),
             );
           }
         },
@@ -929,15 +930,16 @@ function createAppModel({
       const { data, params, highLevelParams, contexts } = processData(
         model.getState()?.rawData as ISequence<IMetricTrace>[],
       );
+      const sortedParams = params.concat(highLevelParams).sort();
       const groupingSelectOptions = [
         ...getGroupingSelectOptions({
-          params: params.concat(highLevelParams).sort(),
+          params: sortedParams,
           contexts,
         }),
       ];
       tooltipData = getTooltipData({
         processedData: data,
-        paramKeys: params,
+        paramKeys: sortedParams,
         groupingSelectOptions,
         groupingItems: ['color', 'stroke', 'chart'],
         model,
@@ -1005,15 +1007,16 @@ function createAppModel({
       if (configData) {
         setAggregationEnabled({ model, appName });
       }
+      const sortedParams = params.concat(highLevelParams).sort();
       const groupingSelectOptions = [
         ...getGroupingSelectOptions({
-          params: params.concat(highLevelParams).sort(),
+          params: sortedParams,
           contexts,
         }),
       ];
       tooltipData = getTooltipData({
         processedData: data,
-        paramKeys: params,
+        paramKeys: sortedParams,
         groupingSelectOptions,
         groupingItems: ['color', 'stroke', 'chart'],
         model,
@@ -2081,7 +2084,7 @@ function createAppModel({
               if (ex.name === 'AbortError') {
                 // Abort Error
               } else {
-                console.log('Unhandled error: ');
+                console.log('Unhandled error: ', ex);
               }
             }
           },
@@ -3025,6 +3028,7 @@ function createAppModel({
                 metric: '',
                 context: [],
                 children: [],
+                groups: groupConfigData,
               };
 
               rows[groupKey!] = {
@@ -3152,7 +3156,7 @@ function createAppModel({
                 rows[groupKey!].data,
                 {},
                 true,
-                Object.keys(columnsValues),
+                ['groups'].concat(Object.keys(columnsValues)),
               );
             }
           },
@@ -3235,6 +3239,7 @@ function createAppModel({
                     }
                   },
                 );
+
                 return {
                   values,
                   color: color ?? run.color,
@@ -3300,16 +3305,16 @@ function createAppModel({
       ): void {
         const { data, params, highLevelParams, metricsColumns } =
           processData(rawData);
-
+        const sortedParams = params.concat(highLevelParams).sort();
         const groupingSelectOptions = [
           ...getGroupingSelectOptions({
-            params: params.concat(highLevelParams).sort(),
+            params: sortedParams,
           }),
         ];
 
         tooltipData = getTooltipData({
           processedData: data,
-          paramKeys: params,
+          paramKeys: sortedParams,
           groupingSelectOptions,
           groupingItems: ['color', 'stroke', 'chart'],
           model,
@@ -3682,14 +3687,15 @@ function createAppModel({
         const { data, params, highLevelParams, metricsColumns } = processData(
           model.getState()?.rawData as IRun<IParamTrace>[],
         );
+        const sortedParams = params.concat(highLevelParams).sort();
         const groupingSelectOptions = [
           ...getGroupingSelectOptions({
-            params: params.concat(highLevelParams).sort(),
+            params: sortedParams,
           }),
         ];
         tooltipData = getTooltipData({
           processedData: data,
-          paramKeys: params,
+          paramKeys: sortedParams,
           groupingSelectOptions,
           groupingItems: ['color', 'stroke', 'chart'],
           model,