+++ CHANGELOG.md
@@ -3,7 +3,11 @@
 ## Unreleased
 
 - Add ability to cancel pending request (roubkar)
+- Add support for secure protocol for API calls (mihran113, roubkar)
 - Implement suggestions list for AimQL search (arsengit, rubenaprikyan)
+- Implement image full size view (VkoHov)
+- Add ability to manipulate with image size and rendering type (arsengit)
+- Enhance Table column for selected grouping config options (arsengit)
 
 ## 3.1.1 Nov 25 2021